third_party/WebKit/LayoutTests/fast/css/sticky/sticky-position-valid-after-layout.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<style>
body {
    margin: 0;
}

#scroller {
    height: 100px;
    width: 200px;
    overflow: scroll;
}

.padding {
    width: 400px;
    height: 500px;
}

#sticky {
    width: 25px;
    height: 25px;
    position: sticky;
    top: 20px;
    left: 50px;
}
</style>

<div id="scroller">
  <div id="sticky"></div>
  <div class="padding"></div>
</div>

<script>
if (window.internals) {
  internals.settings.setCSSStickyPositionEnabled(true);
}

test(function() {
    var scroller = document.getElementById('scroller');
    var sticky = document.getElementById('sticky');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;
    scroller.append(document.createElement('div'));
    assert_equals(sticky.getBoundingClientRect().top, 20);
    assert_equals(sticky.getBoundingClientRect().left, 50);
}, "getBoundingClientRect should be correct for sticky after script-caused layout");

test(function() {
    var scroller = document.getElementById('scroller');
    var sticky = document.getElementById('sticky');
    scroller.scrollTop = 100;
    scroller.scrollLeft = 75;
    scroller.append(document.createElement('div'));
    assert_equals(sticky.offsetTop, scroller.scrollTop + 20);
    assert_equals(sticky.offsetLeft, scroller.scrollLeft + 50);
}, "offsetTop/offsetLeft should be correct for sticky after script-caused layout");
</script>
